Move vulkan to common/video.
