+++ common/video/vulkan/slang_helpers.cpp
