+++ common/video/vulkan/slang_helpers.hpp
