+++ common/video/vulkan/vulkan_hpp_storage.cpp
