+++ common/video/vulkan/vulkan_pipeline_image.cpp
