+++ common/video/vulkan/vulkan_simple_output.cpp
