+++ common/video/vulkan/vulkan_texture.hpp
